requirements.txt
flask~=2.0.2
imdbpy
youtube-search-python
flask_socketio
Flask-Mail
flask-mail
mysql-connector-python~=8.0.27
cachetools~=4.2.4
pyyaml~=6.0
itsdangerous~=2.0.1
Werkzeug~=2.0.2